+++ packages/toolkit/plugin-v2/src/types/cli/plugin.ts
@@ -1,0 +1,8 @@
+import type { Plugin } from '../plugin';
+import type { CLIPluginAPI } from './api';
+/**
+ * The type of the CLI plugin object.
+ */
+export type CLIPlugin<Config, NormalizedConfig> = Plugin<
+  CLIPluginAPI<Config, NormalizedConfig>
+>;