+++ packages/cli/plugin-swc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/core
 
+## 2.62.1
+
+### Patch Changes
+
+- @modern-js/utils@2.62.1
+
 ## 2.62.0
 
 ### Patch Changes