+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/generator-utils
 
+## 3.7.5
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.62.1
+- @modern-js/generator-common@3.7.5
+
 ## 3.7.4
 
 ### Patch Changes