+++ packages/toolkit/e2e/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/e2e
 
+## 2.62.1
+
+### Patch Changes
+
+- @modern-js/utils@2.62.1
+
 ## 2.62.0
 
 ### Patch Changes